requirements.txt
Flask-SQLAlchemy>=1.0
nose==1.3.7
email_validator==1.0.3
iso3166==0.9
pytz==2018.6
isbnlib==3.9.3
py-moneyed==0.7.0
schwifty===2018.9.1